move options to network package
