+++ network/controller/rpc_controller.go
@@ -69,7 +69,6 @@ import (
 	"github.com/insolar/insolar/instrumentation/instracer"
 	"github.com/insolar/insolar/network"
 	"github.com/insolar/insolar/network/cascade"
-	"github.com/insolar/insolar/network/controller/common"
 	"github.com/insolar/insolar/network/hostnetwork/packet/types"
 )
 
@@ -89,7 +88,7 @@ type rpcController struct {
 	Network        network.HostNetwork                `inject:""`
 	OriginProvider network.OriginProvider             `inject:""`
 
-	options     *common.Options
+	options     *network.Options
 	methodTable map[string]insolar.RemoteProcedure
 }
 
@@ -331,6 +330,6 @@ func (rpc *rpcController) Init(ctx context.Context) error {
 	return nil
 }
 
-func NewRPCController(options *common.Options) RPCController {
+func NewRPCController(options *network.Options) RPCController {
 	return &rpcController{options: options, methodTable: make(map[string]insolar.RemoteProcedure)}
 }
