+++ network/gateway/base.go
@@ -57,7 +57,6 @@ import (
 	"github.com/insolar/insolar/network/consensus"
 	"github.com/insolar/insolar/network/consensus/adapters"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
-	"github.com/insolar/insolar/network/controller/common"
 	"github.com/insolar/insolar/network/rules"
 	"github.com/insolar/insolar/network/storage"
 
@@ -98,7 +97,7 @@ type Base struct {
 	ConsensusController   consensus.Controller
 	ConsensusPulseHandler network.PulseHandler
 
-	Options *common.Options
+	Options *network.Options
 
 	bootstrapETA    time.Duration
 	originCandidate *adapters.Candidate
