+++ network/options.go
@@ -48,7 +48,7 @@
 //    whether it competes with the products or services of Insolar Technologies GmbH.
 //
 
-package common
+package network
 
 import (
 	"time"
